requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
coverage==7.8.0
Django==5.1.6
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dock==0.0.1
docx==0.2.4
et_xmlfile==2.0.0
exception==0.1.0
fake-useragent==2.1.0
idna==3.10
lxml==5.3.2
mongoengine==0.29.1
openpyxl==3.1.5
pika==1.3.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pymongo==3.11.4
python-docx==1.1.2
pytz==2025.1
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.3.0
